dynamo_update/requirements.txt
boto3
requests
python-dotenv
tqdm